+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change history
 
+## 09-04-2024
+- added Django 5 support
+- removed dependency django-more-admin-filters
+- fixed type validation for multi-type configs
+- disabled editor, depending on the value type, for multi-type configs (now it works for one-type configs only)
+- updated README
+
 ## 16-02-2024
 - added Python 3.12 support
 - added new example of validator with name decorator